dbench: simplify open_loadfile() as check_loadfile_ok()
open_loadfile() just checks to see if the file.gz exists
and is present. Note that if the file is not in the gzip format,
gzopen and gzread will not produce an error but just read it
without doing any uncompressing. So just rename it to
check_loadfile_ok().

While at it fix the incorrect use of gzFile as a pointer.
Using it as a pointer works just because the file descriptor
can cast to the pointer, but this generates compilation warnings.
Fix that as well.

And lastly, just use gzclose(f) for correctness;

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>